Basic deep zoom functionality now working
Leaflet script now accepts variables on initialization, so it can work
with any image in the collection. Currently the data needs to be
recorded in the site-wide YML data file; from here it is stashed as a
data-attribute for .object__contents on each page. Leaflet reads
attributes from this DOM element on initialization. Perhaps a better
approach will be discovered in the future, but for now this works.