Integrate custom map controls into Leaflet.js map
Previously the fullscreen toggle control was a custom UI element living
in HAML templates for the document. This was less than ideal for a
number of reasons.

This change improves the situation by using a Leaflet plugin,
EasyButton
(http://danielmontague.com/projects/easyButton.js/v1/examples/). Now the
full-screen controls appear as a native leaflet control that will live
alongside other map control elements.

In order to get styling right, a JS override stylesheet had to be added
to the project. This should only be used when absolutely necessary.