Improved Panel Animation
This commit is a work-in-progress attempt to create a choreographed
series of meaningful animated transitions between views in the app. Now
that Smoothstate is working properly, the bulk of the work to get this
running will rely on velocity.js.

The current animation has illusion of panels opening and closing as the
user moves between grid and detail views.

A rendering bug in the first attempt led to a significant removal of
code that was no longer necessary. The panel layout has become greatly
simplified – map panel no longer needs a .panel__header div at all
(since all map/image controls are now generated by Leaflet itself).
After ditching flex elements in the right panel header and using a
simpler layout, the bug has disappeared.