Begin refactoring view layer code
This commit begins a major refactoring of template code in the project.
The goal is to make HAML templates much more simple/legible. A helper
method of `object_data` has been added to config.rb, enabling the
removal of a lot of logic from the object template. This should be the
model for similar refactors going forward. JS files and CSS also changed
in a few places to work with the new HTML structure.