Enable Lunr Search
This commit adds a basic implementation of Lunr client-side search. In
addition to the Lunr library itself, the UI to display results uses a
Handlebars template to render things on the fly. A special
contents.json.erb template has also been added – this generates a
full-text index of hte site on build.