Fix error in mapCheck function
Search controls were not working because of an error in the function
which checks to see if a map exists on the current page. Cause of
problem: when using jQuery to check if an element exists, the code
if($el) {...} will always return true. That's because jQuery returns an
empty array even if no elements are found.

To fix this, make sure to check for $el.length in all places.