Use Velocity.js to scroll to anchor links
This seemingly trivial feature hid a number of JS "gotchas". First,
jQuery has issues selecting an element that contains a colon, so these
have to be escaped. See the jq function inside ui-functions.js for
details.

Secondly, Velocity.js can scroll elements inside of containers with
positions other than "static" but the container needs to be passed in as
an option specifically. So in the scroll function for footnotes a
conditional statement checks to see if the right-panel container exists;
if so, that is passed as an option; if not, the scroll method is called
normally.