Improve panel layout
This commit includes a number of changes – mainly, the layout of panel
elements has been greatly simplified through the use of Flexbox (as
opposed to floats, etc). This fixes several bugs, including a bug where
the page header would disappear whenever the navigation menu was opened.

Additionally, the page header element now animates along with other
elements when the left panel expands. This is a big UI improvement.