Responsive Cover layout
This commit makes several improvements to the layout of the cover page
(and a few other global places like page-header). Cover page now has a
fully responsive layout where the focal point of the cover image is
never covered by text. This is achieved by taking the cover-inner div
and un-nesting it from the .cover div; On medium screens and larger it
is positioned absolutely, on top of the full-height cover (always 50vh
or smaller, and anchored to the bottom half of the screen).

Responsive layout here allows for adjustment of the site-wide
breakpoints for media queries. The medium-screen breakpoint now covers
sizes all the way down to about 600px.

Multiple different background images for the cover are used, one for
each breakpoint. In some cases these images have been cropped so that
nothing important is obscured when the page is rendered.

Page header needs an explicit display:fixed on non-panel layouts; this
has been added as well. Page header is transparent and minimal on the
cover page for the time being.

Finally, now that elements are not crashing in to one another when user
zooms due to weird positioning bugs, this commit has re-enabled user
scaling on touch devices. This is always more desirable for
accessibility purposes.