Update to Leaflet 1.0 beta-1
This commit updates the map-functions.js code to be compatible with the
latest version of Leaflet (1.0 beta 1).

This commit takes a major step towards fixing Issue #10, where images
with differently-sized aspect ratios would partially overlap one
another.

Previously the default/main view of each object was added via an
explicit call to L.tileLayer() outside of the rest of the basemaps
object. Leaflet Control was aware of basemaps but not the additional
default layouer; this probably was the cause of that layers failure to
disappear.

After these changes, all layers (including the main view) are included
inside the basemaps object. The layers: option which is passed on map
initialization is where the default layer is now specified.

Unfortunately, there is a new bug that is yet to be solved here
– non-existent map tiles (outside of the image boundaries) now appear as
broken image icons... Will attempt to fix this next.