Fix broken image icons for most cases
In reference to issue #10. After some experimentiation it looks like the
best way to work with multi-layer image maps is to do things in the
following order:

- Initialize an empty map object with the correct projection
- Create a mapBounds variable using the new L.LatLngBounds()
  constructor. Use map.unproject inside this for each corner.
- Call map.fitBounds(mapBounds)
- Generate a baseMaps object for each view the object has. Specify an
  option of bounds: mapBounds for each layer. All of this happens inside
  the L.tileLayer constructor.
- Add map controls to the map, passing baseMaps as an argument.
- call map.addLayer(default layer) to set the default view.

One remaining issue is the broken image tiles still appear for layers
with a different aspect ratio from the default. One solution to this
would be to set a pixelWidth and pixelHeight for each object that
encompasses the widest/tallest images that object has. This would be
done on a per-object basis. The only downside of this would be some
extra margins around certain views.