Fix Issue #10
This commit fixes Issue #10, where Leaflet was behaving badly when the
layers for an image had different dimensions. It is still useful to have
the dimensions for the default view immediately available (so that
Leaflet can fit the map to the view when it loads). But all views for
all objects now live in the img_index.yml data file. Ideally this could
be generated programatically in the future with something like
ImageMagick.