Refactor next/prev keyboard nav functions
Previously this was done by forcibly changing the window.href value. The
downside of this technique was that if forced a complete page reload
every time, even if smoothstate was active. By calling .click() on the
links instead, Smoothstate will still handle the page transition for
users of keyboard nav.