Add GeoJSON file for map
This is a complete GeoJSON map file. Generated with the AWMC’s
Antiquity À-la-carte application, and further tweaked by hand. Have
added modern place names as they should appear on map by default to the
“custom_name” field. Added the “map_level” and “in_area” fields to
distinguish groups and types of features specific to the catalogue,
based on Claire’s original list. Will still need Claire to determine
which versions of the ancient place names to use.