Enable basic multi-filter grid
The grid component has been totally reworked. Now the catalogue grid is
dynamically generated at runtime using a Handlebars template.

A global gridSelection object manages the state of the catalogue view.
Key/value pairs are added to this object as properties when the user
chooses a category from one of the dropdown menus. So by selecting
"Region: Sicily" on the menu, the gridSelection.region property is
created. It's value would be "sicily".

Each time this happens, the renderGrid function is called. The data
passed to it is a filtered subset of the catalogue entries that match
all the selection criteria (using underscore's _.where() method to do
this). This data is passed to the Handlebars template and re-rendered.

TODO: handle empty states (add a message when no items are returned);
also, adding some animation or loading indicator to indicate that things
have changed would be helpful.