Add "Downloads" and "Cite this Page" to Sidebar
This commit adds in expandable sections in the nav sidebar for download
links as well as "cite this page" functionality. Citation format should
be reviewed for accuracy.

The largest part of this update is the _cite.html.haml file in the
/partials/ folder. The template includes a large conditional statement
to determine the formatting of the citation text, and then provides it
in both MLA and Chicago style. I considered putting this code into a
helper method but the amount of complicated string formatting required
seemed more fitting in a template.

Finally, MLA style requires inclusion of the date accessed. Because
Javascript's built-in Date object doesn't support named months, the
moment.js library has been included to handle this.