Enable global next/previous navigation (#30)
The running head now contains < and > links which work across all major
pages and sections of the publications. New helper methods (next_entry,
prev_entry, next_page, prev_page) have been added to deal with the
underlying logic.

The page header template determines which method is called on a given
page, and the methods output the appropriate Haml tags with a link to
the next item in the sort order.

Catalogue entries do not need any metadata added (their Cat number is
used for sorting). Content pages need to have a sort_order attribute
included in the frontmatter to be included in the sequence. This value
should be a positive integer starting at 0 (currently the cover == 0).