Fix <smallcaps> formatting with Regex transforma
This commit is the result of running a regular expression script on the
contents of the data/catalogue directory. The script runs as follows:

pattern = /(<\/span>)(.*\d{4})(,)/

files.each do |file_name|
  text = File.read(file_name)
  new_text = text.gsub(pattern, '\2\1\3')
  File.open(file_name, "w") { |file| file.puts new_text }
end

Essentially it swaps the order of the closing </span> tag and the
4-digit number after it, leaving the quotation in place. There seem to
be exceptions when a line break character separated the contents of the
span tag, but this got a lot of the errors automatically. Leaving
instructions here in case I have to do this again.