Citation format fixes
Fixed some of the citation formatting. Notably Chicago uses Getty
Publications and MLA uses Getty P, so broke the variables out in finer
grain in the book.yml and the _cite.html.haml files.

Adjusted the indenting on the more info link, so it would show up not
just in the All other pages part of the if statement

Made some related formatting tweaks to the citation info on the about
page, and included the cite-current-date function there as well.