Add Print Template for catalogue pages
A stripped down catalogue page template should be used for generating
catalogue pages in alternate formats (PDF, POD, EPUB, etc.). This
template removes the various JS components and replaces it with an <img>
tag for the main object image. Other images could be added as well,
perhaps by calling the "views" attribute of the entry hash.

This template also removes the group discussion – the idea is that all
entries will appear together in a paged format, so including the
discussion on each one would be a big waste of space.

The master plan would be to create a custom PDF extension that uses the
Sitemap object to proxy the catalogue content to an alternate template
when a "middleman PDF" command is given. When this happens, the
discussion texts should be rendered with the "Page" layout instead of
layout: false.