Add new and revised download files
First commit of the JSOn data files and the PPT download, with
apologies to Eric.

Revised JPG ZIP file. Standardized all images to be 1440px on longest
side, with a density of 72ppi; updated and corrected the caption files.

Corrected Extended CSV to eliminate an empty column on the end.

Updated About page downloads section and download call out in the side
nag to reflect new file and image sizes.