VueJS: Graceful loading behavior w/o DOM querying
Previously the catalogue grid Vue instance would allow users to glimpse
the naked template for a card item before it had completed
loading/rendering the full image set. Users could see things like {{
item.info.title }} where an actual title should be, etc. Initially the
only obvious way to keep this from happening was to manually add/remove
classes via jQuery in the component's "ready" method.

Obviously, this felt like the dirty hack that it was. The whole point of
using something like Vue is to avoid this kind of direct DOM
querying/manipulation in the first place.

It turns out a more graceful solution for this problem is already
provided by the framework.

- For loading indicator elements: instead of manually removing/hiding
	these elements when the component is ready, use a `v-if` directive.
	For example, the Vue instance can contain a data object of "loaded:
	false". When it finishes loading, set it to `true` in the ready
	method. In the respective template, add a `v-if="! loaded"` directive
	to the pre-loader; this means it only remains on the page as long as
	the Vue instance's `loaded` attribute is `false`.

- For hiding pre-compiled template elements: Just put a `v-cloak`
	attribute on the element. That's it!

This commit makes the rendering process for the catalogue grid 100% free
of any direct DOM manipulation.