Basic Vue.js catalogue implementation
At this point a working catalogue component exists as a full-featured
Vue instance. The component contains a range of features from the
framework, including custom filters, directives, and computed
properties. 2-way data binding between the <select> elements on the page
and the catalogue dataset removes the need for any jQuery-based handling
of user input.

One big advantage of Vue is how easy it makes defining various states in
a template. In this example, a computed property called currentEntries
calcultes the result from the currently selected filters. If the length
of this result is zero, a v-if directive switches on with a message for
the user.

One current downside: v-model bindings only work on HTML form elements.
Which means the custom dropdowns from the jQuery version of this page
may be harder to re-create.

One bit of remaining functionality to implement is the "clear filters"
button - this should call a custom method that resets the `selection`
object to its initial state.