Include discussions in client-side search index
Previously discussion pages had been left out of the search index
(created from the contents.json file). However search is a lot less
useful with that text omitted. This commit changes the way the search
index JSON file is generated so that discussion pages are included.
Since they don't live as stand-alone pages, they point to the URL of the
first entry included in the discussed set. To ensure that this behavior
functions consistently, metadata for the discussion pages has been
changed to include a ruby Range object of the catalogue entries.