Fix scoping leak of [data-catalogue] style rules
When creating a sass for-loop to write custom bg-image rules based on
the data-catalogue attribute, I made the mistake of leaving it in the
global scope. There are other elements (mainly the map backgrounds for
some images) which also have this data attribute applied; they were
getting strange styling as a result. This commit fixes this problem by
scoping all [data-catalogue] selectors to .card objects.