Link from Cat. location text to map point (#35)
It turns out there is a leaflet plug-in which makes this pretty easy to
do, Leaflet-hash: http://mlevans.github.io/leaflet-hash/

This commit enables leaflet-hash for the full-screen map, not for other
maps (i.e. the home page version). The URL updates as you scroll around,
and it is now possible to link to a specific zoom level and set of
coordinates. The map loads on its initial view and then quickly zooms in
to the desired location.

Right now the data that matches up lat/lon coordinates to specific
location strings ("Taranto Region", etc.) is hard-coded into a helper
method. This should probably be moved to a data file at some point.