Improve mobile appearance/behavior
This commit includes some style improvements which hide certain features
that don't fit/work on smaller devices, to streamline the experience.

The fullscreen toggle behavior doesn't work properly on iPhone (and
isn't really relevant there in the first place), so this commit includes
a modification of the custom Leaflet button to include an ID of
"#fullscreen-toggle". This button is removed from mobile displays via
CSS.

Additionally, an improved media query has been added for tablets to
detect orientation. When the tablet is in portrait orientation, a
condensed version of the page header is shown in the narrow space. When
orientation changes, this information expands.