Add guide images and initial figures yaml file
Adds 31 images to be used in the Guide. All have been resized to have a
maximum dimension of 1600px. The YAML file only includes captions, for
now, for the 3 images that were taken from an iPad. The rest only have
the source file name and an arbitrary id until we figure out how we’ll
handle them.