Add PDF Output extension
This commit adds a basic Middleman extension for PDF output. The basic
process works as follows:
- On `middleman build`, a new version of each catalogue page is created,
  using the `print-template.html` file. This is proxied to the
  `/print-catalogue/` directory.
- The extension loops through all pages wiht a `sort_order` defined and
  creates a list of contents in the correct reading order: frontmatter,
  catalogue entries, backmatter. See the extension source for details.
- This list is fed to the Prince CLI (the extension assumes it is
  available), which generates a single PDF file of the entire book. A
  print CSS stylesheet helps make the result look better.

Future work: Right now the reliance on an external dependency (Prince)
is not very well documented. Some way of checking/requiring that this is
available would be useful to add. Also tests.

This plugin could probably be easily adapted to work for Epub output as
well (though in that case a set of template files ought to be included
in the extension, so they do not clutter up the rest of the app).