Improved print version generation
This commit enables a fully-working print version, automatically
generated with PrinceXML and a PDF extension. Several layout adjustments
had to be made to show/hide content where appropriate.

The largest change concerns the Table of Contents. In order to generate
page numbers and working document links with Prince, URLs in the TOC
(for cross-referencing) must match those which are supplied to the
Prince command-line program. This required building out a separate
version of the TOC in-place, which only appears inside a print media
query.