Improve page header styles (plus webkit bug fix)
This commit uses css line-height to consistently style the height of
page header elements (icons, etc). This works better than using
absolute/relative "top" positioning.

Additionally, this commit fixes a very frustrating Mobile/Desktop Safari
bug where the search results would sometimes just disappear while the
user was typing a query. Apparently there is a known bug where Safari
does not re-paint the page when it should in certain circumstances where
a fixed position element is being added/changed on the page.

The solution to this was to force a repaint in Javascript. When the code
runs which triggers the search results to appear, I've added the
following bit of code:

$('<style></style>').appendTo($(document.body)).remove();

After adding this code, I can no longer reproduce the bug on desktop or
mobile Safari.