Add trailing slash on internal links
Because every page is in its own directory and we’re hijacking links to
give us better transitions, page links break when the page loads from a
URL without the trailing slash. For example, clicking to the About page
from an internal link structured as “../about” worked b but resulted in
links on the About page being improperly formatted and so broken.
Changing the internal link to “/about/“ fixes the issue.