Enable epub export (working with caveats)
This commit enables a basic epub workflow through Pandoc. This is still
somewhat of a hack and is about 75% there. A new template has been added
for epub layout of catalogue entries (where the JS-dependent catalogue
layout is not appropriate). Like when generating a PDF, there is a line
in config.rb which proxies all catalogue entries to this template and
outputs them into an /epub/ directory. This should probably be deleted
after building; there is no need for these files to be publicly
available.

A number of epub-specific assets have also been added to the /assets/
folder, in Fonts (open source fonts safe for embedding), Images, and
Stylesheets. Note that fonts and stylesheets must be manually included
via a flag when running pandoc.

To actually create an epub file, spin up the server and run the pandoc
script in the /bin folder. This hits all the main content areas (Guide
needs to be added) and creates an epub with this content.

Some manual optimizations then need to be performed: for example,
manually wiring up the TOC and nav elements to look for anchor links in
the single HTML file that constitutes the epub content. Hopefully this
can be improved in the future. However, the output file does work fine
in iBooks.