Fix strange paragarph formatting issue
In the list for the Guide to the Collection, some items are not
rendering as paragraphs, leaving their font size slightly smaller than
normal. In most cases this is fixed by adding a hard line break in
between the list items.

For a few entries this still didn't produce the expected results (often
the last item in a section). Where necessary I've wrapped these elements
in <p> tags manually.