Leaflet button fixes
This commit fixes one issue and temporarily hides another.
The rotation icon was not showing up properly, so some style overrides
have been applied to the .easy-button-button class to force proper
display.

A second issue is that on ipad/mobile views, the fullscreen mode is not
working properly any longer. For the time being I have simply disabled
this feature from all touchscreen devices by adding a display:none rule
to the fullscreen button when touch is detected. This rule can be
removed when we discover a fix.