Update epub and mobi files
While the last version of these ebook files were successfully validated
with epubcheck version 4.0.1, one or more of our vendors uses the older
epubcheck 3.0.1 to validate files and was receiving an error on all
footnote link IDs that used a colon (such as, fn:2 and fnref:10). Here,
 the colon was simply replaced with a dash throughout the book (like,
fn-2 and fnref-10).